Moved CmdLine back to LogExpert as it is not part of the core logic of the application.
